Convert all PlayerMoveEvents to tasks, increasing CPU efficiency
