+++ .gitignore
@@ -1,6 +1,25 @@
-.gitignore
+.DS_Store*
+ehthumbs.db
+Icon?
+Thumbs.db
+bin
+target
+*.glob
+.*.glob
+.classpath
+.project
+.settings
 .idea/
-Elytra.iml
-META-INF/
-src/test/
-target/
+*.iml
+*.ipr
+*.iws
+dependency-reduced-pom.xml
+/target/
+*.class
+.project
+.classpath
+/.settings/
+
+/lib/
+.settings
+.git
