+++ pom.xml
@@ -12,6 +12,10 @@
 			<id>spigot-repo</id>
 			<url>https://hub.spigotmc.org/nexus/content/repositories/snapshots/</url>
 		</repository>
+		<repository>
+        	<id>DrkMatr1984 Repo</id>
+        	<url>http://minevolt.net/repo</url>
+    	</repository>
 	</repositories>
 	<dependencies>
 		<dependency>
@@ -26,7 +30,7 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>RELEASE</version>
+				<version>3.5</version>
 				<configuration>
 					<source>1.8</source>
 					<target>1.8</target>
